Write tests for CustomerJDBCDataAccessService.java, refactoring Integer -> Long

+++ src/main/java/com/borjas/customer/CustomerController.java
@@ -17,7 +17,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{customerId}")
-    public Customer getCustomer(@PathVariable("customerId") Integer customerId) {
+    public Customer getCustomer(@PathVariable("customerId") Long customerId) {
         return customerService.getCustomer(customerId);
     }
 
@@ -27,12 +27,12 @@ public class CustomerController {
     }
 
     @DeleteMapping("/{customerId}")
-    public void deleteCustomer(@PathVariable("customerId") Integer customerId) {
+    public void deleteCustomer(@PathVariable("customerId") Long customerId) {
         customerService.deleteCustomerById(customerId);
     }
 
     @PutMapping("/{customerId}")
-    public void updateCustomer(@PathVariable("customerId") Integer customerId,
+    public void updateCustomer(@PathVariable("customerId") Long customerId,
                                @RequestBody CustomerUpdateRequest updateRequest) {
         customerService.updateCustomer(customerId, updateRequest);
     }
