+++ src/main/java/com/borjas/customer/CustomerJDBCDataAccessService.java
@@ -28,7 +28,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public Optional<Customer> selectCustomerById(Integer customerId) {
+    public Optional<Customer> selectCustomerById(Long customerId) {
         var sql = """
                 SELECT id, name, email, age
                 FROM customer
@@ -65,7 +65,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public boolean existsPersonWithId(Integer id) {
+    public boolean existsPersonWithId(Long id) {
         var sql = """
                 SELECT count(id)
                 FROM customer
@@ -77,7 +77,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public void deleteCustomerById(Integer customerId) {
+    public void deleteCustomerById(Long customerId) {
         var sql = """
                 DELETE
                 FROM customer
