+++ src/main/java/com/borjas/customer/CustomerRepository.java
@@ -3,7 +3,7 @@ package com.borjas.customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
     boolean existsCustomerByEmail(String email);
-    boolean existsCustomerById(Integer id);
+    boolean existsCustomerById(Long id);
 }
